Core

Documentation

updateWorld :: Float -> World -> World

Updates a world state by a single time step. Internally it uses a Reader monad to convey read-only state (timestep, environment variables) to the individual updates. | Note that the simulation is timestep-dependent, so 2 updates of 1.0 won't give the same result as 1 update of 2.0. updateWorld :: Float -- ^ The time step to update by -> World -- ^ The world to update -> World

updateSquare :: Square -> Reader WorldState Square

Creates a Reader for updating the square; you would use this reader by creating a WorldState and doing runReader (updateSquare square) worldState.

newPop :: GoodPrices -> SquarePop -> Reader WorldState SquarePop

produceGoods :: SquarePop -> GoodQuantities

calcSupplies :: GoodQuantities -> GoodTrades -> GoodTrades -> GoodQuantities

calcTradeVolumes :: GoodTrades -> GoodTrades -> GoodQuantities

calcLocalUtilities :: SquarePop -> GoodQuantities -> GoodUtilities

newPrices :: GoodQuantities -> GoodQuantities -> GoodUtilities -> SquareNeighbours -> GoodPrices -> Reader WorldState GoodPrices

newTradeIn :: SquareRef -> SquareNeighbours -> GoodTrades

newTradeOut :: GoodQuantities -> GoodPrices -> SquareNeighbours -> GoodTrades -> Reader WorldState GoodTrades

flattenTrade :: GoodTrades -> GoodTrades -> GoodTrades

findGdef :: a -> Good -> [(Good, a)] -> a

findG0 :: Good -> [(Good, Float)] -> Float

sumTradeGood :: Good -> GoodTrades -> Float

goodList :: GoodTrades -> [Good]

findSquare :: World -> SquareRef -> Square

findNTup :: [(Neighbour, b)] -> (Int, Int) -> (Neighbour, b)

findNeigh :: SquareNeighbours -> SquareRef -> Neighbour

findNConn :: SquareNeighbours -> SquareRef -> Float

makeNeighbour :: Square -> Neighbour

paramLK :: WorldState -> String -> Float